chore: fix traffic lint and update agents guide
